+++ app/models/like.rb
@@ -19,6 +19,8 @@ class Like < Federated::Relayable
 
   after_destroy do
     self.parent.update_likes_counter
+    participation = author.participations.where(target_id: target.id).first
+    participation.unparticipate! if participation.present?
   end
 
   xml_attr :positive